WinnerLottery, WinnerLotteryTest  * WinnerLottery 추상 객체 추가 및 테스트 코드 추가
